backend/requirements.txt
fastapi
psycopg2-binary
requests
shapely
geojson
uvicorn